Coarse Filter class implementation and Ball improvments
