+++ Foosball2text/Ball.cs
@@ -14,13 +14,14 @@ namespace Foosball2text
     {
         private float x;
         private float y;
-        private Point center;
+        private CircleF _circle;
         public Ball()
         {
         }
 
         public float X { get => x; set => x = value; }
         public float Y { get => y; set => y = value; }
+        public CircleF Circle { get => _circle; set => _circle = value; }
 
         private CircleF[] GetCirclesFromFrame(Image<Gray, byte> frame)
         {
@@ -37,17 +38,14 @@ namespace Foosball2text
 
         public void FindCordinates(Image<Gray, byte> frame)
         {
-            foreach (CircleF circle in GetCirclesFromFrame(frame))
+            CircleF[] circleArray = GetCirclesFromFrame(frame);
+            if (circleArray.Length != 0)
             {
-                x = circle.Center.X;
-                y = circle.Center.Y;
+                Circle = circleArray[circleArray.Length-1];
+                X = _circle.Center.X;
+                Y = _circle.Center.Y;
             }
         }
-        public void Draw()
-        {
-
-        }
-
     }
 
 }
